requirements.txt
# Core application dependencies
fastapi
python-dotenv
openai
pydantic
python-multipart

# Langchain - using specific packages to reduce size
langchain-core
langchain-openai
langchain-community

# PDF processing
pypdf

# Other tools
python-jobspy
